FIX: error when deal with empty body request
Previously, the request was read up to the header, emptied the buffer
and set the write_fdset.
However, if the buffer is empty before selection, the read_fdset is
emptied, so there was a problem that the write_fdset was not set if
there was no buffer to empty at the time of reading the request to the
header.