FEAT: check redirection functions
Implement reason:
If location block is set as below in config file and acces to
'old-folder' route, it must be redirected to 'new-folder' route.
```
locatino /old-folder {
    return 301 /new-folder/;
}
```
Therefore, it must be checked if 'return' key is in location info.
And findRedirectionStatusCode would return status code which is belong
to 'return' key.